+++ src/DependencyInjection/Configuration.php
@@ -6,6 +6,8 @@ namespace Dbp\Relay\FormalizeBundle\DependencyInjection;
 
 use Dbp\Relay\BlobLibrary\Api\BlobApi;
 use Dbp\Relay\BlobLibrary\Api\BlobApiError;
+use Dbp\Relay\CoreBundle\Authorization\AuthorizationConfigDefinition;
+use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -13,6 +15,7 @@ class Configuration implements ConfigurationInterface
 {
     public const DATABASE_URL = 'database_url';
     public const DEFAULT_BLOB_TYPE = 'default_blob_type';
+    public const MAY_ADD_FORM = 'MAY_ADD_FORM';
 
     /**
      * @throws BlobApiError
@@ -30,7 +33,16 @@ class Configuration implements ConfigurationInterface
              ->end();
 
         $rootNode->append(BlobApi::getConfigNodeDefinition());
+        $rootNode->append($this->getAuthorizationNodeDefinition());
 
         return $treeBuilder;
+    }
+
+    private function getAuthorizationNodeDefinition(): NodeDefinition
+    {
+        return AuthorizationConfigDefinition::create()
+            ->addResourcePermission(self::MAY_ADD_FORM, 'false',
+                'Returns true if the user has allowed to add the given form, false otherwise')
+            ->getNodeDefinition();
     }
 }