+++ src/DependencyInjection/DbpRelayFormalizeExtension.php
@@ -7,6 +7,7 @@ namespace Dbp\Relay\FormalizeBundle\DependencyInjection;
 use Dbp\Relay\CoreBundle\Doctrine\DateTimeImmutableUtcType;
 use Dbp\Relay\CoreBundle\Doctrine\DoctrineConfiguration;
 use Dbp\Relay\CoreBundle\Extension\ExtensionTrait;
+use Dbp\Relay\FormalizeBundle\Authorization\AuthorizationService;
 use Dbp\Relay\FormalizeBundle\Service\SubmittedFileService;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -29,8 +30,10 @@ class DbpRelayFormalizeExtension extends ConfigurableExtension implements Prepen
         );
         $loader->load('services.yaml');
 
-        $definition = $container->getDefinition(SubmittedFileService::class);
-        $definition->addMethodCall('setConfig', [$mergedConfig]);
+        $container->getDefinition(SubmittedFileService::class)
+            ->addMethodCall('setConfig', [$mergedConfig]);
+        $container->getDefinition(AuthorizationService::class)
+            ->addMethodCall('setConfig', [$mergedConfig]);
 
         $this->addResourceClassDirectory($container, __DIR__.'/../Entity');